CREP-11 Fix login form so it accepts credentials and reroutes to homepage
