+++ src/contexts/CurrentUserContext.js
@@ -19,7 +19,7 @@ export const CurrentUserProvider = ({ children }) => {
 			const { data } = await axiosRes.get("dj-rest-auth/user/");
 			setCurrentUser(data);
 		} catch (err) {
-			// console.log(err);
+			console.log(err);
 		}
 	};
 
@@ -36,7 +36,7 @@ export const CurrentUserProvider = ({ children }) => {
 					} catch (error) {
 						setCurrentUser((prevCurrentUser) => {
 							if (prevCurrentUser) {
-								history.push("/login");
+								router.push("/login");
 							}
 							return null;
 						});
@@ -60,7 +60,7 @@ export const CurrentUserProvider = ({ children }) => {
 					} catch (err) {
 						setCurrentUser((prevCurrentUser) => {
 							if (prevCurrentUser) {
-								history.push("/login");
+								router.push("/login");
 							}
 							return null;
 						});
