+++ src/hooks/useRedirect.js
@@ -1,26 +1,26 @@
 import axios from "axios";
+import { useRouter } from "next/router";
 import { useEffect } from "react";
-import { useHistory } from "react-router-dom";
 
 export const useRedirect = (userAuthStatus) => {
-    const history = useHistory();
+	const router = useRouter();
 
-    useEffect(() => {
-        const handleMount = async () => {
-            try {
-                await axios.post("/dj-rest-auth/token/refresh/");
-                // if user is logged in, the code below will run
-                if (userAuthStatus === "loggedIn") {
-                    history.push("/");
-                }
-            } catch (err) {
-                // if user is not logged it, the code below will run
-                if (userAuthStatus === "loggedOut") {
-                    history.push("/");
-                }
-            }
-        };
+	useEffect(() => {
+		const handleMount = async () => {
+			try {
+				await axios.post("/dj-rest-auth/token/refresh/");
+				// if user is logged in, the code below will run
+				if (userAuthStatus === "loggedIn") {
+					router.push("/");
+				}
+			} catch (err) {
+				// if user is not logged it, the code below will run
+				if (userAuthStatus === "loggedOut") {
+					router.push("/");
+				}
+			}
+		};
 
-        handleMount();
-    }, [history, userAuthStatus]);
+		handleMount();
+	}, [router, userAuthStatus]);
 };
